antlr4cpp on Windows
- Solved all compilation issues. Updated the antlr4cpp library project. At the moment we only build a static lib. Need to add exports for a DLL. Since we want this library to be compatible with VS 2013 still, we cannot use std::rethrow_with_nested(), hence we do a simple unnested throw in such cases. Starting with VS 2015 this works fully then.
- Added demo application.
- Added parser generation script.

+++ runtime/Cpp/runtime/Parser.h
@@ -355,7 +355,7 @@ namespace runtime {
 
     /// <summary>
     /// Get a rule's index (i.e., {@code RULE_ruleName} field) or -1 if not found. </summary>
-    virtual int getRuleIndex(const std::wstring &ruleName);
+    virtual ssize_t getRuleIndex(const std::wstring &ruleName);
 
     virtual ParserRuleContext::Ref getRuleContext();
