+++ runtime/Cpp/runtime/support/CPPUtils.h
@@ -33,7 +33,7 @@
 namespace antlrcpp {
 
   std::wstring join(std::vector<std::wstring> strings, const std::wstring &separator);
-  std::map<std::wstring, int> toMap(const std::vector<std::wstring> &keys);
+  std::map<std::wstring, size_t> toMap(const std::vector<std::wstring> &keys);
   std::wstring escapeWhitespace(std::wstring str, bool escapeSpaces);
   std::wstring toHexString(const int t);
   std::wstring arrayToString(const std::vector<std::wstring> &data);
